+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.5.1] - 2025-02-15
+### Fixed
+- TBA queue detection
+
 ## [0.5.0] - 2024-12-24
 ### Removed
 - All CLI configuration arguments - Use the configuration file or environmental vars in future
@@ -78,7 +82,8 @@ env var. For Docker builds, this will be `/config/settings.json`
 
 Initial release
 
-[Unreleased]: https://github.com/CPU-Blanc/scrubarr/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/CPU-Blanc/scrubarr/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/CPU-Blanc/scrubarr/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/CPU-Blanc/scrubarr/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/CPU-Blanc/scrubarr/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/CPU-Blanc/scrubarr/compare/v0.3.0...v0.3.1